refactor(auth.dto): move access token response dto
